Updating podspec and fixing lint issue

+++ Pod/Classes/UITableViewController+IGTableViewEstimatedHeightCache.m
@@ -57,7 +57,7 @@
 #pragma mark - Helpers
 
 - (NSString *)cacheKeyForIndexPath:(NSIndexPath *)indexPath {
-    return [NSString stringWithFormat:@"%ld-%ld", indexPath.section, indexPath.row];
+    return [NSString stringWithFormat:@"%ld-%ld", (long) indexPath.section, (long) indexPath.row];
 }
 
 @end
